+++ src/core/SftpFileSystem.cpp
@@ -754,8 +754,8 @@ public:
 
     if (Flags & SSH_FILEXFER_ATTR_EXTENDED)
     {
-      unsigned int ExtendedCount = GetCardinal();
-      for (intptr_t Index = 0; Index < ExtendedCount; ++Index)
+      unsigned long ExtendedCount = GetCardinal();
+      for (intptr_t Index = 0; Index < static_cast<intptr_t>(ExtendedCount); ++Index)
       {
         GetRawByteString(); // skip extended_type
         GetRawByteString(); // skip extended_data
@@ -838,7 +838,7 @@ public:
   UnicodeString Dump() const
   {
     UnicodeString Result;
-    for (intptr_t Index = 0; Index < GetLength(); ++Index)
+    for (intptr_t Index = 0; Index < static_cast<intptr_t>(GetLength()); ++Index)
     {
       Result += ByteToHex(GetData()[Index]) + L",";
       if (((Index + 1) % 25) == 0)
@@ -3140,7 +3140,7 @@ void TSFTPFileSystem::ReadDirectory(TRemoteFileList * FileList)
         unsigned int Count = ListingPacket.GetCardinal();
 
         TRACE("4");
-        for (intptr_t Index = 0; !isEOF && (Index < Count); ++Index)
+        for (intptr_t Index = 0; !isEOF && (Index < static_cast<intptr_t>(Count)); ++Index)
         {
           File = LoadFile(&ListingPacket, NULL, L"", FileList);
           if (FTerminal->GetConfiguration()->GetActualLogProtocol() >= 1)
